requirements.txt
flask
spidev
RPi.GPIO
numpy
Pillow
gpiozero
lgpio